# internal/server/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Bongsu Vulnerability Management API
  version: 0.1.0
  description: >
    Vulnerability management and SBOM platform API. Provides host inventory,
    vulnerability scanning, CVE database management, RBAC, and installer
    endpoints.
  license:
    name: MIT
servers:
  - url: /
    description: Current server
security: []
paths:
...
  /api/intelligence/recommendations:
    get:
      operationId: getRecommendations
      tags: [intelligence]
      summary: Get remediation recommendations
      security:
        - AdminKey: []
        - ViewerKey: []
        - bearerAuth: []
      responses:
        "200":
          description: List of recommendations
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Recommendation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/admin/notification-rules:
    get:
      operationId: listNotificationRules
      tags: [notifications]
      summary: List notification rules
      security:
        - AdminKey: []
        - bearerAuth: []
      responses:
        "200":
          description: List of notification rules
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/NotificationRule"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: createNotificationRule
      tags: [notifications]
      summary: Create a notification rule
      security:
        - AdminKey: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationRuleCreate"
      responses:
        "201":
          description: Notification rule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationRule"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/admin/notification-rules/{id}:
    get:
      operationId: getNotificationRule
      tags: [notifications]
      summary: Get a notification rule
      security:
        - AdminKey: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notification rule details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationRule"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      operationId: updateNotificationRule
      tags: [notifications]
      summary: Update a notification rule
      security:
        - AdminKey: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationRuleUpdate"
      responses:
        "200":
          description: Notification rule updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationRule"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteNotificationRule
      tags: [notifications]
      summary: Delete a notification rule
      security:
        - AdminKey: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notification rule deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/admin/notification-log:
    get:
      operationId: listNotificationLog
      tags: [notifications]
      summary: List notification delivery log
      security:
        - AdminKey: []
        - bearerAuth: []
      parameters:
        - name: rule_id
          in: query
          schema:
            type: string
          description: Filter by notification rule ID
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Paginated notification log entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/NotificationLog"
                  total:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Phase 4d: Reports ──────────────────────────────────────────────────
  /api/reports/executive-summary:
    get:
      operationId: getExecutiveSummary
      tags: [reports]
      summary: Generate executive summary report
      security:
        - AdminKey: []
        - ViewerKey: []
        - bearerAuth: []
      responses:
        "200":
          description: Executive summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutiveSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/reports/risk-breakdown:
    get:
      operationId: getRiskBreakdown
      tags: [reports]
      summary: Get risk breakdown by group
      security:
        - AdminKey: []
        - ViewerKey: []
        - bearerAuth: []
      parameters:
        - name: group_by
          in: query
          schema:
            type: string
          description: "Metadata field to group by (owner, team, environment, criticality)"
      responses:
        "200":
          description: Risk breakdown rows
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/RiskBreakdownRow"
                  group_by:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
...
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key

    bearerAuth:
      type: http
      scheme: bearer
      description: Session token

    AdminKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Admin API key

    AgentKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Agent API key (or admin key)

    ViewerKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Viewer API key (subject:key format)

    InstallToken:
      type: apiKey
      in: header
      name: X-Install-Token
      description: Install token

  # ── Reusable Parameters ──────────────────────────────────────────────
  parameters:
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 50
      description: Maximum items to return
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        maximum: 1000000
        default: 0
      description: Number of items to skip

  # ── Reusable Responses ───────────────────────────────────────────────
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      content:
        text/plain:
          schema:
            type: string
            example: unauthorized
    Forbidden:
      description: Forbidden
      content:
        text/plain:
          schema:
            type: string
            example: forbidden
    NotFound:
      description: Not found
      content:
        text/plain:
          schema:
            type: string
            example: not found
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    AtRiskHost:
      type: object
      properties:
        host_id:
          type: string
        hostname:
          type: string
        total_vulns:
          type: integer
        critical_count:
          type: integer
        high_count:
          type: integer
        exploited_count:
          type: integer
        overdue_count:
          type: integer
        max_risk_score:
          type: number
          format: double
    Recommendation:
      type: object
      properties:
        priority:
          type: integer
        category:
          type: string
        title:
          type: string
        description:
          type: string
        host_ids:
          type: array
          items:
            type: string
...
    NotificationRule:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        enabled:
          type: boolean
        trigger_event:
          type: string
        min_severity:
          type: string
        min_risk_level:
          type: string
        exploited_only:
          type: boolean
        host_filter:
          type: string
        channel_type:
          type: string
        channel_config:
          type: object
          additionalProperties: true
        last_triggered:
          type: string
          format: date-time
          nullable: true
        trigger_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    NotificationRuleCreate:
      type: object
      required: [name, trigger_event]
      properties:
        name:
          type: string
        enabled:
          type: boolean
          default: true
        trigger_event:
          type: string
        min_severity:
          type: string
        min_risk_level:
          type: string
        exploited_only:
          type: boolean
          default: false
        host_filter:
          type: string
        channel_type:
          type: string
        channel_config:
          type: object
          additionalProperties: true
    NotificationRuleUpdate:
      type: object
      properties:
        name:
          type: string
        enabled:
          type: boolean
        trigger_event:
          type: string
        min_severity:
          type: string
        min_risk_level:
          type: string
        exploited_only:
          type: boolean
        host_filter:
          type: string
        channel_type:
          type: string
        channel_config:
          type: object
          additionalProperties: true
    NotificationLog:
      type: object
      properties:
        id:
          type: string
        rule_id:
          type: string
        trigger_event:
          type: string
        status:
          type: string
        payload:
          type: object
          additionalProperties: true
        error_message:
          type: string
          nullable: true
        attempts:
          type: integer
        created_at:
          type: string
          format: date-time

    # ── Phase 4d: Reports ───────────────────────────────────────────────
    ExecutiveSummary:
      type: object
      properties:
        generated_at:
          type: string
          format: date-time
        total_hosts:
          type: integer
        host_coverage_percent:
          type: number
          format: double
        active_vulnerabilities:
          type: integer
        severity_counts:
          type: object
          additionalProperties:
            type: integer
        risk_level_counts:
          type: object
          additionalProperties:
            type: integer
        exploited_count:
          type: integer
        overdue_sla_count:
          type: integer
        sla_compliance_percent:
          type: number
          format: double
        trend_direction:
          type: string
          enum: [up, down, flat]
        trend_delta:
          type: integer
        top_risk_hosts:
          type: array
          items:
            $ref: "#/components/schemas/AtRiskHost"
    RiskBreakdownRow:
      type: object
      properties:
        group:
          type: string
        total:
          type: integer
        severity_counts:
          type: object
          additionalProperties:
            type: integer
        risk_level_counts:
          type: object
          additionalProperties:
            type: integer
...